+++ CotEditor/Sources/Document Window/Text View/EditorTextView+Indenting.swift
@@ -91,6 +91,7 @@ extension EditorTextView: Indenting {
     
     var tabWidth: Int { get }
     var isAutomaticTabExpansionEnabled: Bool { get }
+    var indentTokens: [IndentToken] { get }
 }
 
 
@@ -132,7 +133,7 @@ extension Indenting {
         guard
             self.tabWidth > 0,
             let selectedRanges = self.rangesForUserTextChange?.map(\.rangeValue),
-            let textEditing = self.string.smartIndent(style: self.indentStyle, indentWidth: self.tabWidth, in: selectedRanges)
+            let textEditing = self.string.smartIndent(style: self.indentStyle, indentWidth: self.tabWidth, tokens: self.indentTokens, in: selectedRanges)
         else { return false }
         
         return self.edit(with: textEditing)