+++ CHANGELOG.md
@@ -26,6 +26,7 @@
     - TypeScript
 - Prepend a symbolic icon to extracted outline items.
 - Make outline items in the Outline inspector collapsible when using a tree-sitter-based syntax.
+- Introduce syntax-based smart indentation.
 - Enable registering multiple comment delimiters in a syntax definition.
 - Add the Multiline option to syntax highlighting rules.
 - Add the Lexical Rules pane to the syntax editor to customize parsing rules.